+++ src/constants/panel.ts
@@ -5,3 +5,11 @@ export const HeaderOptions = [
   { value: true, label: 'Display', description: 'Display variable Label or Id.' },
   { value: false, label: 'Hide', description: 'Hide header.' },
 ];
+
+/**
+ * Sticky Position
+ */
+export const StickyOptions = [
+  { value: true, label: 'Enabled', description: 'Follow when scrolling.' },
+  { value: false, label: 'Disabled', description: 'Disable sticky position.' },
+];